FastAPI refactor of survos2.
