+++ survos2/_dock_widget.py
@@ -8,14 +8,13 @@ import survos2
 import sys
 import warnings
 from loguru import logger
+from survos2.frontend.control import Launcher
 
 warnings.filterwarnings("ignore")
-
+from survos2.server.state import cfg
 
 # fmt = "{time} - {name} - {level} - {message}"
-fmt = (
-    "<level>{level} - {message} </level><green> - {name}</green>:<cyan>{function}</cyan>:<cyan>{line}</cyan>"
-)
+fmt = "<level>{level} - {message} </level><green> - {name}</green>:<cyan>{function}</cyan>:<cyan>{line}</cyan>"
 logger.remove()  # remove default logger
 # logger.add(sys.stderr, level="DEBUG", format=fmt, colorize=True)
 logger.add(sys.stderr, level=Config["logging.overall_level"], format=fmt, colorize=True)
